added

+++ src/main/java/com/example/springdataprojects/service/EmployeeServiceImpl.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 
 @Service
-public class EmployeeServiceImpl implements EmployeeService{
+public class EmployeeServiceImpl implements EmployeeService {
 
     private final EmployeeRepository employeeRepository;
 
@@ -23,24 +23,31 @@ public class EmployeeServiceImpl implements EmployeeService{
     }
 
     @Override
-    public List<Employee> getAllEmployes() {
+    public List<Employee> getAllEmployees() {
         return employeeRepository.findAll();
     }
 
     @Override
-    public Employee addEmployee(Employee employee) {
+    public boolean addEmployee(Employee employee) {
         employeeRepository.save(employee);
-        return employee;
+        return true;
     }
 
     @Override
-    public Employee updateEmployee(Employee employee) {
-        employeeRepository.save(employee);
-        return employee;
+    public boolean updateEmployee(Employee employee) {
+        if (employeeRepository.existsById(employee.getId())) {
+            employeeRepository.save(employee);
+            return true;
+        }
+        return false;
     }
 
     @Override
-    public void deleteEmployee(Long id) {
-        employeeRepository.deleteById(id);
+    public boolean deleteEmployee(Long id) {
+        if (employeeRepository.existsById(id)) {
+            employeeRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
